refactor(menu): move _generate_config_actions from options menu to base menu
